feat: adding clear page example where a dialog can be fired (from parent component); using useImperativeHandle, forwardRef, Ref
